// src/types/api/status/shipment-status.ts
export type ApiShipmentStatus =
  | 'created'
  | 'offers_prepared'
  | 'offer_selected'
  | 'confirmed'
  | 'dispatched_by_sender'
  | 'dispatched_by_sender_to_pok'
  | 'collected_from_sender'
  | 'taken_by_courier_from_pok'
  | 'taken_by_courier'
  | 'adopted_at_source_branch'
  | 'sent_from_source_branch'
  | 'adopted_at_sorting_center'
  | 'sent_from_sorting_center'
  | 'adopted_at_target_branch'
  | 'delay_in_delivery'
  | 'ready_to_pickup'
  | 'ready_to_pickup_from_pok'
  | 'out_for_delivery'
  | 'out_for_delivery_to_address'
  | 'delivered'
  | 'returned_to_sender'
  | 'avizo'
  | 'undelivered_wrong_address'
  | 'undelivered_incomplete_address'
  | 'undelivered_unknown_receiver'
  | 'rejected_by_receiver'
  | 'claimed'
  | 'cancelled';